+++ src/main/java/szewek/mcflux/R.java
@@ -8,13 +8,11 @@ public final class R {
 			MF_FULL_NAME = "Minecraft-Flux",
 			MF_URL = "https://minecraft.curseforge.com/projects/minecraft-flux",
 			MF_DEPS = "required-after:fl",
-			MF_PKG = "szewek.mcflux",
-			MF_FLC_KEY = "EaIaD392qp12ZLzxNb2PpSewadmHuSQZf4UsK03JVmM",
 			MF_ACCESS_TOKEN = "65a919ed8de14b019f9f51200635c4d6",
 			MF_ENVIRONMENT = "production",
-			GUI_FACTORY = MF_PKG + "szewek.mcflux.gui.MCFluxGuiFactory",
-			PROXY_SERVER = MF_PKG + "szewek.mcflux.proxy.ProxyCommon",
-			PROXY_CLIENT = MF_PKG + "szewek.mcflux.proxy.ProxyClient";
+			GUI_FACTORY = "szewek.mcflux.gui.MCFluxGuiFactory",
+			PROXY_SERVER = "szewek.mcflux.proxy.ProxyCommon",
+			PROXY_CLIENT = "szewek.mcflux.proxy.ProxyClient";
 
 	public static final int MF_GUI_FLUXGEN = 1;